+++ exercisely-frontend/package.json
@@ -11,7 +11,12 @@
     "react-dom": "^18.1.0",
     "react-scripts": "5.0.1",
     "react-webcam": "^7.0.1",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "@mediapipe/pose": "^0.4.1633558788",
+    "@tensorflow-models/pose-detection": "^0.0.6",
+    "@tensorflow/tfjs-backend-webgl": "^3.8.0",
+    "@tensorflow/tfjs-converter": "^3.8.0",
+    "@tensorflow/tfjs-core": "^3.8.0"
   },
   "scripts": {
     "start": "react-scripts start",
