+++ exercisely-frontend/src/Exercise.css
@@ -11,7 +11,7 @@
   height: 480px;
 } */
 
-/* #theCanvas {
+#theCanvas {
   position: absolute;
   left: 0;
   top: 0;
@@ -20,4 +20,4 @@
   width: 640px;
   height: 480px;
   background-color: rgba(255, 0, 0, 0.166);
-} */
+}
